mcp-perf-suite/streamlit-ui/requirements.txt
# MCP Performance Suite - Streamlit UI Dependencies

# Core Framework
streamlit>=1.46.0
streamlit-autorefresh>=1.0.1
python-dotenv>=1.0.1

# YAML Processing (ruamel preserves comments on round-trip edit)
ruamel.yaml>=0.18.0
pyyaml>=6.0.2

# Data Processing & Visualization
pandas>=2.2.2
numpy>=1.26.4
altair>=5.0.0

# Config Migration (diff computation)
deepdiff>=8.0.0

# Image Processing (for file formats e.g. PNG, JPEG)
pillow>=12.0.0